+++ ChangeLog.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## 0.4.1
+## 0.4.3 2024-07-01
+- fix prep_logging to return logger
+
+## 0.4.2 2024-06-29
+- adjust the logging color of cli
+
+## 0.4.1 2024-06-29
 - rename cli: generic-system to import-generic-system
 - add cli: export generic system - WIP
 - implement cli: version option
+